requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.0
manim==0.18.0
manim-slides==3.0.0
manim-voiceover==0.3.1
tavily-python==0.2.6
llama-index-core==0.10.1
llama-index-llms-azure-openai==0.1.1
pydantic==2.6.1 